refactor: recoil reset 코드를 직접 사용하는 곳으로 내려줌.
