+++ pages/index.tsx
@@ -6,42 +6,14 @@ import MainReview from "components/main/MainReview";
 import MainToProcessButton from "components/main/MainToProcessButton";
 import Share from "components/main/Share";
 import type { NextPage } from "next";
-import { useResetRecoilState } from "recoil";
-import {
-  currentCategoryState,
-  currentSkillState,
-  currentSortingState,
-  isDisableState,
-  isOpenState,
-  isSelectedState,
-  lectureDataList,
-} from "store/state";
 
 const Home: NextPage = () => {
-  const resetLectureListData = useResetRecoilState(lectureDataList);
-  const resetCurrentSorting = useResetRecoilState(currentSortingState);
-  const resetIsDisable = useResetRecoilState(isDisableState);
-  const resetIsOpen = useResetRecoilState(isOpenState);
-  const resetIsSelected = useResetRecoilState(isSelectedState);
-  const resetCurrentCategoryData = useResetRecoilState(currentCategoryState);
-  const resetCurrentSkillData = useResetRecoilState(currentSkillState);
-
-  const resetData = () => {
-    resetCurrentCategoryData();
-    resetCurrentSkillData();
-    resetLectureListData();
-    resetIsDisable();
-    resetIsOpen();
-    resetCurrentSorting();
-    resetIsSelected();
-  };
-
   return (
     <div>
       <SEO />
       <Banner />
-      <MainLectureCategory resetData={resetData} />
-      <MainToProcessButton resetData={resetData} />
+      <MainLectureCategory />
+      <MainToProcessButton />
       <MainReport />
       <MainReview />
       <Share />
